+++ CHANGELOG.md
@@ -7,11 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### Fixes
+
+- Restored the markdown-to-pretext command.
+
 ## [0.25.2] - 2025-03-28
 
 ### Fixes
 
-- Fixed bug prevent formatter from working.
+- Fixed bug prevent formatter from working (but it's still not working).
 
 ## [0.25.1] - 2025-03-28